Moved the diff popup into its own file
